+++ onnx/src/ops/array/slice.rs
@@ -58,8 +58,8 @@ impl InferenceRulesOp for Slice {
         s.equals(&inputs.len, 1)?;
         s.equals(&outputs.len, 1)?;
         if self.axes.is_none() {
-            s.equals(&inputs[0].rank, self.starts.len() as i64)?;
-            s.equals(&inputs[0].rank, self.ends.len() as i64)?;
+            s.equals(&inputs[0].rank, self.starts.len() as i32)?;
+            s.equals(&inputs[0].rank, self.ends.len() as i32)?;
         }
         s.equals(&inputs[0].rank, &outputs[0].rank)?;
         s.equals(&inputs[0].datum_type, &outputs[0].datum_type)?;
@@ -77,10 +77,10 @@ impl InferenceRulesOp for Slice {
                 };
                 if let Some((mut b, mut e)) = spec {
                     if let Ok(d) = d.to_integer() {
-                        if b as i64 > d {
+                        if b as i32 > d {
                             b = (d as isize).into();
                         }
-                        if e as i64 > d {
+                        if e as i32 > d {
                             e = (d as isize).into();
                         }
                     }
